refactor(server): migrate controllers to latest standards
