+++ server/src/app/http/controller.ts
@@ -1,0 +1,6 @@
+import express from 'express'
+
+export type Controller = (
+  request: express.Request,
+  response: express.Response,
+) => Promise<void>
