+++ server/src/app/http/index.ts
@@ -1,2 +1,3 @@
+export * from './controller'
 export { default as registerErrorHandler } from './errorHandler'
 export { default as registerRoutes } from './routes'
