+++ server/src/app/modules/classSchedules/list/index.ts
@@ -1,0 +1,1 @@
+export { default as listClassSchedulesController } from './listClassSchedules.controller'
