+++ server/src/app/modules/classSchedules/routes.ts
@@ -1,12 +1,11 @@
 import express from 'express'
 
-import ClassController from '../../../controllers/ClassController'
+import { listClassSchedulesController } from './list'
 
 function register(prefix: string, appRouter: express.Router) {
   const router = express.Router()
-  const classCtrl = new ClassController()
 
-  router.get('/', classCtrl.index)
+  router.get('/', listClassSchedulesController)
 
   appRouter.use(prefix, router)
 }
