+++ server/src/app/modules/classes/create/index.ts
@@ -1,0 +1,1 @@
+export { default as createClassController } from './createClass.controller'
