+++ server/src/app/modules/classes/routes.ts
@@ -1,12 +1,11 @@
 import express from 'express'
 
-import ClassController from '../../../controllers/ClassController'
+import { createClassController } from './create'
 
 function register(prefix: string, appRouter: express.Router) {
   const router = express.Router()
-  const classCtrl = new ClassController()
 
-  router.post('/', classCtrl.store)
+  router.post('/', createClassController)
 
   appRouter.use(prefix, router)
 }
