+++ server/src/app/modules/connections/create/createConnection.controller.ts
@@ -1,0 +1,14 @@
+import HttpStatus from 'http-status-codes'
+
+import database from '../../../../database/connection'
+import { Controller } from '../../../http'
+
+const createConnectionController: Controller = async (request, response) => {
+  const { user: user_id } = request.body
+
+  await database('connections').insert({ user_id })
+
+  response.status(HttpStatus.CREATED).send()
+}
+
+export default createConnectionController
