+++ server/src/app/modules/connections/getCount/index.ts
@@ -1,0 +1,1 @@
+export { default as getConnectionsCountController } from './getConnectionsCount.controller'
